feat: apply connect-v2 updates (#20)
* apply connect-v2 updates

* update readme versions

* change block height from uint64 to int64 & lint (#21)

* change block height from uint64 to int64 & lint

* update readme

* bug checksum length checker

* Update README.md

* bump OPinit to v0.5.1

---------

+++ challenger/host/handler.go
@@ -15,7 +15,7 @@ func (h *Host) beginBlockHandler(_ context.Context, args nodetypes.BeginBlockArg
 }
 
 func (h *Host) endBlockHandler(_ context.Context, args nodetypes.EndBlockArgs) error {
-	blockHeight := uint64(args.Block.Header.Height)
+	blockHeight := args.Block.Header.Height
 	batchKVs := []types.RawKV{
 		h.Node().SyncInfoToRawKV(blockHeight),
 	}
